requirements.txt
psycopg2
requests
pandas
python-dotenv
